# swagger.json
{
    "swagger":"2.0",
    "info": {
        "version": "1.0.0",
        "description": "...SWAGGER DEDCRIPTIN..."
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "schemes": [
        "http"
      ],
      "consumes": [
        "application/json"
      ],
      "produces": [
        "application/json"
      ],
      "paths":{
...
            "/userView":{
              "post":{
                "tags":["View"],
                "description":"view product",
                "parameters":[
                  {
                    "name":"authorization",
                    "in":"header",
                    "required":"true"
                  },
                  {
                    "name":"cart",
                    "in":"body",
                    "description":"view options",
                    "schema":{"$ref":"#/definitions/view"}
                  }
                ],
                "produces":["application/json"],
                "responses":{
                  "200":{"description":"view result "
                  }
                }
              }
            },
            "/cartView":{
              "post":{
                "tags":["View"],
                "description":"cart product",
                "parameters":[
                  {
                    "name":"authorization",
                    "in":"header",
                    "required":"true"
                  },
                  {
                    "name":"cart",
                    "in":"body",
                    "description":"cart options",
                    "schema":{"$ref":"#/definitions/view"}
                  }
                ],
                "produces":["application/json"],
                "responses":{
                  "200":{"description":"cart result "
                  }
                }
              }
            },
            "/adminView":{
              "post":{
                "tags":["Admin View"],
                "description":"category + product",
                "parameters":[
                  {
                    "name":"authorization",
                    "in":"header",
                    "required":"true"
                  },
                  {
                    "name":"cart",
                    "in":"body",
                    "description":"product options",
                    "schema":{"$ref":"#/definitions/adminView"}
                  }
                ],
                "produces":["application/json"],
                "responses":{
                  "200":{"description":"product result "
                  }
                }
              }
            }
    },
    "definitions":{
...
        "view":{
          "required":["CategoryLimit","productLimit","skip"],
          "properties":{
            "CategoryLimit":{
              "type":"number"
            },
            "productLimit":{"type":"number"},
            "skip":{"type":"number"}
          }
        },
        "adminView":{
          "required":["CategoryLimit","CategorySkip","productID","productLimit","productSkip"],
          "properties":{
            "CategoryLimit":{"type":"number"},
            "CategorySkip":{"type":"number"},
            "productLimit":{"type":"number"},
            "productID":{"type":"string"},
            "productSkip":{"type":"number"}
          }
        }
    }
}